+++ src/tutorial/src/step-10/App/options.js
@@ -1,0 +1,20 @@
+export default {
+  data() {
+    return {
+      todoId: 1,
+      todoData: null
+    }
+  },
+  methods: {
+    async fetchData() {
+      this.todoData = null
+      const res = await fetch(
+        `https://jsonplaceholder.typicode.com/todos/${this.todoId}`
+      )
+      this.todoData = await res.json()
+    }
+  },
+  mounted() {
+    this.fetchData()
+  }
+}
