+++ src/tutorial/src/step-10/App/template.html
@@ -1,0 +1,4 @@
+<p>Todo id: {{ todoId }}</p>
+<button @click="todoId++">Fetch next todo</button>
+<p v-if="!todoData">Loading...</p>
+<pre v-else>{{ todoData }}</pre>
