+++ src/tutorial/src/step-9/App/composition.js
@@ -2,22 +2,10 @@ import { ref } from 'vue'
 
 export default {
   setup() {
-    const todoId = ref(1)
-    const todoData = ref(null)
-
-    async function fetchData() {
-      todoData.value = null
-      const res = await fetch(
-        `https://jsonplaceholder.typicode.com/todos/${todoId.value}`
-      )
-      todoData.value = await res.json()
-    }
-
-    fetchData()
+    const p = ref(null)
 
     return {
-      todoId,
-      todoData
+      p
     }
   }
 }
