+++ src/tutorial/src/step-9/App/template.html
@@ -1,4 +1,1 @@
-<p>Todo id: {{ todoId }}</p>
-<button @click="todoId++">Fetch next todo</button>
-<p v-if="!todoData">Loading...</p>
-<pre v-else>{{ todoData }}</pre>
+<p ref="p">hello</p>
