+++ src/tutorial/src/step-9/_hint/App/options.js
@@ -1,25 +1,5 @@
 export default {
-  data() {
-    return {
-      todoId: 1,
-      todoData: null
-    }
-  },
-  methods: {
-    async fetchData() {
-      this.todoData = null
-      const res = await fetch(
-        `https://jsonplaceholder.typicode.com/todos/${this.todoId}`
-      )
-      this.todoData = await res.json()
-    }
-  },
-  created() {
-    this.fetchData()
-  },
-  watch: {
-    todoId() {
-      this.fetchData()
-    }
+  mounted() {
+    this.$refs.p.textContent = 'mounted!'
   }
 }
